Adds support for dragging between subviews.
When placing TKDragView objects in different children (eg. scrollview
with objects) set the parentContainer property to the common ancestor.
It will be used to transform positions through the hierarchies.

This change required updating TKCGRectValue into a real object with a
parent UIView property so that different good rectangles can be in
different children.